Cell-fits-font: integer pointSize + Monaco-natural cell metrics
Picks up iTerm2's playbook for the xterm font path. FontResolver now
snaps pointSize to the nearest integer where Monaco fits both requested
cell dimensions, then reports Monaco's actual advance + lineHeight as
the cell rather than the named-alias dimensions. The XLFD's `7x14`
(etc.) becomes a hint of intended size, not a contract.

Reported metrics === rendered metrics by construction (both come from
the same CTFont at the same integer pointSize), so the "feels bold"
residue from fractional-pointSize-loses-Core-Text-hinting is gone.
xterm's `-fn 7x14` now reports as 6×13 (Monaco at 10pt) and renders
crisply at 3× scale. See XTERM_FONT_QUALITY.md for the empirical
alias map.

Also rolls in the fractional-scale plumbing from prior in-flight work:
DisplayConfig.scale Int→Double propagated through ServerConfig,
FlippedXView, CocoaWindowBridge; font-smoothing-off in ImageText8 /
PolyText8 from the 2026-05-08 "feels bold" iteration.

Doc condense: rewrote XTERM_FONT_QUALITY as "how we render xterm now"
(was journey/brief), trimmed CHATGPT_REVIEW to executive summary +
the principle that drove the metrics work, dropped ARTICLE_BRIEF
(article published), updated SERVER_RESOLUTION cell-snapping math to
the actual algorithm, condensed PRODUCT_2 status entries.

DECISIONS.md 2026-05-09 entry records the choice and what it rejected
(forced-cell, asymmetric stretch, per-alias substitution, smart-stretch).
296/296 tests green.

+++ Sources/SwiftXServerCore/ServerConfig.swift
@@ -26,9 +26,12 @@ public struct ServerConfig: Sendable {
     public var heightInMillimeters: UInt16
     public var vendor: [UInt8]
     public var releaseNumber: UInt32
-    /// Integer multiplier from logical pixels to device pixels. The X
-    /// protocol layer never sees this; the rendering layer uses it.
-    public var scaleFactor: Int
+    /// Multiplier from logical pixels to device pixels. The X protocol
+    /// layer never sees this; the rendering layer uses it. Integer values
+    /// (1, 2, 3) are the Phase-1 happy path with clean N×N device-pixel
+    /// blocks per logical pixel; fractional values like 2.5 are accepted
+    /// at the cost of AA edges between cells.
+    public var scaleFactor: Double
 
     /// Studio Display preset. Used as fallback when no real display info
     /// is available (e.g., test environment where the session is driven
@@ -47,7 +50,7 @@ public struct ServerConfig: Sendable {
         heightInMillimeters: UInt16,
         vendor: [UInt8],
         releaseNumber: UInt32,
-        scaleFactor: Int = 1
+        scaleFactor: Double = 1
     ) {
         self.rootWindowId = rootWindowId
         self.defaultColormapId = defaultColormapId
